add plumbing to connect up character comparison; add test
improve the folder for convert on bools

DCE was broken. This adds back the Op traits to make it work correctly.
Also tightens and repairs two tests, resp.

Replace internal symbol map.

The internal symbol map had served its purpose and needs to be replaced with a
mapping structure that is more appropriate for Fortran variables, which can be
composed of a number of runtime values.  This new map will allow us to lower
more complicated entities correctly by tracking their component values.