Reverse conditional create step so that it is created lazily rather than speculatively. This eliminates creation of some empty fir.if blocks.
Fix a bug in argument passing. When the expression is already a reference, we don't need or want to create a reference to the reference to pass the original reference. Instead just pass the reference itself.

add FIXME for POINTER and ALLOCATABLE

tighten some checks
fix type mismatch on assignment
misc fixes for LAPACK
format bullet items