Adjust the C/C++ compiler environments
The hello world sample and the end-to-end character assignment
test both use the host GCC to build and link their executable.
Some environments use GCC with the env vars C_INCLUDE_PATH and
CPLUS_INCLUDE_PATH set; test must be preserved in the lit env.

Until flang gets a proper driver, the end-to-end "hello world"
example uses a pipeline of bbc, tco, llc, and as to create the
test's PIC relocatable object file. On some systems, GCC needs
the -fPIC to override the default choice of a static link.