Passing pointer to functions not defined in the program unit
Add an interface to CallInterface.h to delcareFunction from a ProcedureDesignator.
This allow creating mlir::FuncOp when lowering indirection to the ProcedureDesignator
if this function was never seen before.

On call sites, add a way to handle argument number mismatch. This is now needed to handle
cases where a procedure is first pass with no type information (the created mlir::FuncOp
has no arguments), and a call to it is later lowered with arguments. This is dealt with
by introducing a function type cast on the the calling site. This is not ideal, but the
alternative that would be to analyze all call site before lowering anything requires
more expensive analysis. If the front-end does this at some point to provide error messages
regarding incompatibles calls, we would take advantage of this.