Support for target specific lowering in the Tilikum bridge.
To generate correct code for a chosen target, the Tilikum bridge must know
what the selected target is and the conventions used for the specific target
ABI. The properties of the target influence the calling conventions and LLVM
IR that must be generated. Tilikum is the last point before any high-level
abstractions must be considered and correctly translated to LLVM IR.

These changed rework the Tilikum bridge to use a target specifier and convert
the calling conventions and memory layouts appropriate for the selected target.

Two target specifications are implemented. i386-unknown-linux-gnu and
x86_64-unknown-linux-gnu. Others can be added as needed.

Two high-level type abstractions are considered: COMPLEX and CHARACTER.

Moving these target specific lowerings to a common place in code gen eliminates
the need to perform heroics with custom code in lowering and/or reliance on
assuming the target is known by implication at compiler compile-time.