Do not reflect the size of character in boxchar length
When passing a character array in an implicit interface we were
multiplying the array size with the length to compute the
boxchar length. It gives the actual storage length, but was not
in line with what other compilers do, leading to link time
incompatibility on F77 programs.
We were anyway not dividing back the length for character array dummy
on the callee size so this did not work correctly between flang
compiled programs.

There are internal cases in IO where we still do want to transform contiguous
array into a scalar, add test to avoid regressions when passing array as
format.