Array expression lowering. Part 2.
Lowering of array assignments. This makes the old lowering of array expressions and assignments obsolete.

Expression lowering is bifurcated into the lowering of scalar expressions and array expressions (>= F90). This allows a natural nested scoping of sub-expression contexts as each sub-expression is lowered. The scalar expression code was mostly kept intact with some cleanup. The array expression lowering uses CPS (continuation passing style) to lower array expressions in parts: setup, loop nest and body, and cleanup.

While this is incomplete, it provides an interface and an implementation and an initial implementation for some target cases.