Lowering assumed shape arrays
Handle assumed shape array both on callee/caller sides:

- Modify the call interface lowering utility to handle explicit interface,
  add TODOs for anything but intrinsic type assumed shape arrays.

- On caller side, simply add a call to createBox in case the interface needs
  a box.

- On callee side, the lowering of bounds is done using `fir.box_dims`, the
  address is obtained with `fir.box_addr` and for character the length is
  done using `fir.box_elesize`.

Note that fir generation works as expected, but lowering to LLVM on the callee
side fails due to a bad GEP generated for fir.box_dims.

Assumed shape array do not require any change in the symbol map used in lowering
because their shape and base address cannot change in the program unit (as
opposed to allocatable/pointers that are deferred shape arrays).