Get assumed Shape working end-to-end
- Always reflect rank when lowering fir.box type to its llvm
type (make the dimension array explicit). This is needed to
use the LLVM GEP when lowering fir.box_dims on callee side.

- Fix my own incompetence of getting bad indexes for lower_bound
and extent (1 and 2 instead of 0 and 1).

- The front-end is not making implicit assumed shape lower bounds
explicit (to 1) but is labeling as deferred. Due to this, the
code was taking the lower bound of the descriptor instead of
of 1. Workaround this for now, when we lower allocatable and
pointer, we will want either to update the front-end to make assumed
shape lower bound explicit or to handle this better in lowering.

- Align lit tests

With this patch, assumed shape associated with contiguous intrinsic
variable are compiling OK end to end.