Lower global allocatables
- Emit inline code to lower allocatable initialization and remove
  AllocatableInitIntrinsic runtime usage. This is done by using fir.embox and
  taking advantage of fir.store on boxes.

- Add an initialization for the global fir.box using fir.embox.
Move the allocatable and pointer related code from mapSymbolAndAttribute
into a central place since it is the same regarding of the rank.
Like for other allocatables, the support is limited to non character
intrinsic for now.

- Change type conversion to be the place where the fir.box type is added
for allocatable and pointers instead of doing it afterwards in the bridge
(to better enforce single point of decision for types).

- Disable pointer related tests until pointers are implemented (hard todo
have been added until then).