Lowering modules
All the work done in this patch is around module variables since module
procedure were already mangled OK and have nothing special in lowering,
apart that they can access module data.

In PFT variable:
- add isDeclaration() to indicate if the scope is using a variable that
it does not own (because it is in a module for instance).
- ramp-up the alias analysis to take into account equivalences imported
 from used modules.

In the bridge:
- split instantiateGlobal function into: declareGlobal, defineGlobal,
  instantiateGlobal.
- split instantiateAggregateStore functions into: declareGlobalAggregateStore,
  defineGlobalAggregateStore, and instantiateAggregateStore
- split instantiateCommon into: defineCommon, and, instantiateCommon.

"declareXX" is meant to be used to create global op for a variable that
is defined in another file (not owned by any of the scopes being lowered).
"defineXX" is meant to be used to create global op for a variable defined
in this file.
"instantiateXX" is meant to create a SymbolBox and add it into the
symbol map. It may call declareXX or defineXX if the globalOp needed
to build the symbol address does not exit yet.

Common blocks only have a defineXX method because one cannot say if they
don't belong to any scope. The common linkage ensure this is not an
issue.

I took the opportunity to organize all the code instantiating variables
to make upstream review easier.

Note: a front-end fix will be pushed in LLVM to make the use side of
equivalenced module data not defined in the file working.
Submodules need testing.