Remove GEP usage in embox codegen
Always build a descriptor value using UndefOp + InsertOps.
Then, if the embox is not done in a GlobalOp body, allocate storage
for the box and store the value that was built.

This simplifies the codegen code and the produced LLVM IR. The resulting
assembly (at least on X86 64), is the same.