Array slicing.
Implement the slicing operation on an array. The slice can be over elements or
subcomponents of the contiguous base array. The result is the logical projection
of the original array as a new array. This projection may be stored in a
temporary array.
Implements vector-subscripts in array sectioning/slicing.
Bonus: fir.zero_bits Op.
Changes due to rebasing.