Fix clang11 + libstdc++ 10.2 issue
When returning a variant that has a member with a template ctor,
clang11 attempts to call this ctor even when we really obvisouly want to build
a different members of the variant. This result in a bogus compilation failure.
Use std::in_place to indicate which variant is being built.