Fix local character array lowering
createNewLocal was used to build character arrays without considering
the dynamic lengths.
Upgarde createNewLocal and FirOpBuilder::allocateLocal to take an array of lengths
parameter.
For now, merge the lengths and the shape in allcoateLocal because fir::AllocaOp is
ignoring it lengths arguments (fixme added).
Remove usage of createCharacterTemp in variable lowering (use createNewLocal instead)
to make character processing similar in all cases and to have the variable name in the
allocaOp.