Associate constructs (#572)
Implement associate construct associations where the associating entity
is assigned the value of the selector expression, or is an alias of a
whole selector variable. Associations where the associating entity is
an alias of a subobject of a selector variable are related to pointer
assignments, and are not yet implemented. Generated code for
associations such as "associate (b => a, a => b)" will be incorrect
pending a front end fix.

See F18 Clause 19.5.1.6p3