Add fir.absent and fir.is_present operation
It is not possible to express a fir.box argument as being
null pointer pointer in fir (casts between fir.box and fir.ref are
forbidden). To implement Fortran OPTIONAL argument, it is needed to
create fir.box as null pointer on the caller side and to be able to
check if they are null pointers on the callee side.

To avoid exposing fir.box as being pointer in all situations, create ad-hoc
fir.absent and fir.is_present operations.
fir.absent will create a fir.box that is a null pointer in LLVM.