Add fir.shift op to be used with fir.box
When working with fir.box in fir array operation, the extents are not
added to the operation since they are already in the fir.box. However,
local lower bounds may not be reflected in the fir.box. Add a fir.shift
type and operation to provide lower bounds only to fir array operations
in this case.