[OpenMP] Fixes for worksharing loop lowering
The OpenMP worksharing loop operation in the dialect is a proper loop
operation and not a container of a loop. We have to lower the OpenMP
loop construct and the do-loop inside the construct to a omp.wsloop
operation and there should not be a fir.do_loop inside it. In this fix
this is achieved by skipping fir.do_loop creation and calling genFIR
for the nested evaluations in the lowering of the do construct. When a
worksharing loop is created an index is created for it. All usage of the
index in the loop is replaced with this index.