[mlir] Translate global initializers after creating all LLVM IR globals
In case an operation in a global initializer region refers to another
global variable defined afterwards in the module of itself, translation
to LLVM IR was currently crashing because it could not find the LLVM IR global
when going through the initializer block.

To solve this problem, split global conversion to LLVM IR in two passes. A
first pass that create LLVM IR global variables, and a second one that lower
the initializer, if any, and add it to the llvm global.