Resolve bind names to strings in semantics
Convert binding labels for BIND(C) entities to strings during semantic
processing and save them in the symbol. This takes care of stripping
spaces and the default name when no name is specified. Make it available
through Symbol::GetBindName().

This is the change from https://reviews.llvm.org/D99208 with the
corresponding changes to Lower/CallInterface.cpp to make use of the
new API.