Constant array constructors (#749)
* Constant array constructors

Implement a helper class to generate a unique name for a typed literal constant value based on an (MD5) message digest hash.  Use that when generating .rodata constant array definitions to allow reuse of those values.  This helper class should be valid for all constants - scalars and arrays of intrinsic or derived type if other uses turn out to be useful.  Names are type specific.  If the same bits are defined twice with different types, their hash names have a type component to distinguish between them.

Constants are always accessed directly without referencing a named symbol.  Suppress generation of an unused, dead .rodata definition tied to the symbol that is not needed for correct code generation.  If some connection between the symbol and its value is eventually needed, such when generating debug information, the symbol name will need to be tied to the hash name.  It is possibile to do that, but that will require accessing the constant from a higher expression level, which is a little more difficult than direct access to a low level constant expression.