Lower explicit results depending on symbols
- Add buildFuncResultDependencyList() to the PFT tools so that the
list of pft::Variables that a results depends upon can be obtained at
the call site. As opposed to normal PFT analysis, this list is built on
demand and not saved because the function being called may not have
a pft::Node where to save this (only functions defined in the current
unit have pft::Nodes). It is needed to do such analysis to deal with the
cases where the result depends on module or common block symbols that
could be equivalenced.

- Add mapCallInterfaceSymbols to instantiate the variables that the result
specification expressions depends upon. It is called on the caller side
before evaluating the specification expressions that contains references
to these variables.

- Replace typeAndShape->LEN() usage because it returns an expressions
that depends on the result symbol (descriptor inquiry) instead of the
specification expression.