Finish lowering F95 structure-constructor and derived assignments
Add:
 - Handling pointer components
 - Handling non allocatable/non pointer array components
 - Fixing character handling in structure-component (was generating a
   field coordinate_of on the rhs expr, as if in derived type assignment).

To so:
- Share componentToExtentedValue so that both structure-component and
  assignment code can match the extended value instead of a custom if
  chain.
- Split initializer case for structure-constructor to avoid too many
  branching.
- Expose two new array assignment entry points: One work with both LHS
  and RHS being ExtendedValue (used in assignment, will be useful for
  copy-in/copy-out around function calls), and one where the LHS is an
  ExtendedValue and the RHS is an evaluate::Expr (used in
  structure-constructor). Add templates to lowerArrayAssignment to avoid
  duplicating the constituent semantic set-up.

Note:
It is likely the assignment case could be dealt with a load/store of the
structure, but the component by component approach will be needed when
facing allocatable component or component with user defined assignments.

Allocatable components left with hard TODO. For component that have
user defined assignment, a FIXME is left since I do not know yet how
to detect these case (note that gfortran and ifort do not implement this
yet, though it is pretty clear this is needed from 10.2.1.3 point 13).