Embox array slices in IO item list instead of making copies
The IO used genExprAddr that creates contiguous copies for non contiguous
entities (array slices and assumed shapes). This is semantically wrong in IO
input since the actual variable was not modified as it should. This is also
wasting resources in IO output since the IO runtime support non
contiguous descriptors, so there is no need to create a temp and making
a copy.

- Add AbstractConverter::genExprBox that creates a fir.box describing an
 expression without making a temporary if the expression is a designator.
 It uses createSomeArrayBox on array expressions.

- Modifies the IO input/output list code to use genExprBox when the IO item
 must be passed through a descriptors (which is the case for arrays).

- Update genarr(ArrayRef) embox case to pass the length to EmboxOp when this
  is needed.

- Fix getTypeParams for arrays (do not return character length when
  length is constant but array is dynamic).