Fix indices origin in elemental calls
For elemental call arguments, a fir.array_coor is generated in array expression
lowering. The fir.array_coor operation expects indices based on the entity
origin (1 by default), but array expression indices are zero based.

I considered making the shape argument passed to the fir.array_coor a zero
based shapeshift, but that does not play well with cases where there is a slice
(that needs the actual entity origin in the shape).
So instead, share originateIndices that is doing the same job in the
array copy pass. The code has to be templatized in a header because
the array value and lowering do not use builders of the same type
(FIROpBuilder vs PatternRewritter).