Fix search of unrestricted intrinsics
Lowering of intrinsic passed as arguments was failing because
the runtime search was mixing-up result and argument types (for
most intrinsic, it bug did not matter, but for intrinsic with more than
one argument, like atan2 it was casing issues).

Also enable a TODO test related to dummy procedure for which lowering
has been implemented since then.