do not expect defining op when undoing complex pattern
The previous code was doing an unconditional usage of getDefiningOp
result on the root of the two fir.complex insertOps to check that it was
an undefOp.
This is wrong because it is not guaranteed that the root is not a
block or function argument, especially in case of later mismatch.

This actually caused a bug #806 from Fortran code because some MLIR
optimization pass combined some code operating on complexes into a new
block while leaving the root undefOps in the parent op and passing the
root as block argument (I think it is weird, but I have no idea what
exactly is doing this, and this specific rewrite is hard to trigger,
nearly any modification to #806 code stop MLIR from doing this rewrite).

The pattern rewrite is already checking the result is coming from two
insertValueOp on complex that insert in 0 and 1, hence it actually does not
matter to check that the root is an undefOp. So simply remove this check.

Add a test that operations can still be grouped even if the UndefOp is not
visible. Also add fir tests for these complex pattern rewrites since there was
none yet (there was only tests from Fortran).