Implement most of allocatable assignment
Done:
 - Assignment to allocatable scalar intrinsic, and derived types without
   without length parameters. Deferred length scalar character are covered.
 - Allocatable array assignment to numeric, logical and character with non
   deferred length.

Missing:
 - Support for assignment to allocatable character array with deferred
   length
 - Support for polymorphic allocatable. They are more requirements for
   those.
 - Tests for simple derived type arrays are not yet added, nothing particular
   is needed regarding the allocatable aspect, but it needs to be
   supported in normal array assignment, which is used after the
   conditional reallocation.

How:

Add a genReallocIfNeeded function to Allocatable.cpp that is
centralizing the generation of the conditional reallocation logic both
for the scalar and array assignment cases.

For the scalar path, evaluate rhs first so that length parameters, if
needed, are available to perform the conditional re-allocation.
Perform conditional reallocation, and proceed with the regular
assignment path.

For the array path, add createAllocatableArrayAssignment entry point to
the array framework, that takes a MutableBoxValue for the lhs, and
drive the reallocation and assignment.
The MutableBoxValue is kept as an optional pointer, in the framework,
the rhs is lowered to continuation as usual. When generating the
iteration space, gather the rhs shape, length (TODO), lbounds if any,
and pass that to genReallocIfNeeded. Then, load the allocatable into
arrayLoad and proceed with array assignment as usual.