Host association in internal procedure part 2: non trivial vars
This patch adds lowering for host associated characters, arrays,
pointers, and allocatables.

- Characters are associated by placing a fir.boxchar into the host tuple.
- Arrays are associated by placing a pointer descriptor (fir.box<fir.ptr>)
  inside the host tuples.
- Pointers and allocatables are associated by placing their descriptor address
  (fir.ref<fir.box<>>) inside the host tuple.

A special attentions is given to:
 - propagating array lower bounds. Simple fir.box<T> would lose this
   info, using pointer association with fir.box<fir.ptr<T>> ensures
   these do not get lost.
 - handling host associated optional dummies. The packaging/un-packaging
   of such variable should account for the fact that the dummy cannot be
   unconditionally read. Besides, the optional aspect must be preserved
   inside the internal procedure (PRESENT must work there).
 - Keeping allocatable and pointers non-deferred type parameters knowledge.
   This knowledge has impacts on the semantic of statement where such
   entities can appear inside the host (e.g. allocatable assignment).

To deal with this, the host association handling is refactored to group
the work according to the type of variable instead of having the work
group according to the phases (getting the type, building the tuple,
instantiating the variable). This allows better mirror reading of things
that must happen when packaging and unpacking the tuple for special
kind of variables.