Set extent of impossible triplet slices to zero
The previous lowering/codegen code was not accounting for the fact
that in slice (lb:ub:step), it is legal to have `ub-lb` and `step`
with a different sign, in which case the resulting dimension extent
is zero as per 9.5.3.3.2 point 3 and 4.

This led to negative extents to be computed and used, leading to
crashes like #832.