Use array_fetch on where masks instead of fir.coordinate_of
fir.coordinate_of was used to fetch pre-computed values inside
where statements in a way that was only valid with compile time
constant shape masks.

To fix this, there was three alternatives:
 - Keep fir.coordinate_of, but collapse dimensions an compute offset
   in lowering.
 - Use fir.array_coor, taking into account lower bounds.
 - Use fir.array_load + fir.array_fetch, directly using loop indexes.

I went for the last one that was the easiest to deploy, and adds the
less index computation noises inside the fir loops, which as I understand
should be better for future loop optimizations (affine...).