Fix local character array with lower bounds and static length
A length was passed to fir.alloca for a static length character type
when instantiating StaticArrayStaticChar in variable lowering.

This caused error: 'fir.alloca' op LEN params do not correspond to type

Do not pass the length in this case, and add tests of local with lower
bounds.

Also remove IsAllocatedOrPointer from the asserts in variable lowering
since this case is handled before, and not expected where the assert
were (This was not an issue, but troubled me when fixing this bug).