Lower NULL(MOLD) and function results as pointers
NULL(MOLD) is still represented as an intrinsic call after lowering
(which is not the case for NULL() that is an evaluate::NullPointer).
Lower NULL(MOLD) in the intrinsic framework.

Function returning pointers were already lowered OK, however, the
contexts where the function results is still considered to be a pointer
were not handled yet because there is only one such context: when
passing a result pointer to a pointer arguments. Update genMutableBoxValue
to deal with this case.

To do so, move genPorcedureRef content inside genRawProcedureRef that does not
automatically read pointers/allocatable results into normal values,
and keep this automatic read in genPorcedureRef.

Last, the case foo(NULL()) needs special handling in call lowering because
the created disassociated pointer takes the characteritsics of the dummy,
and cannot be lowered by genMutableBoxValue().