Fix array%character array section not in CharBoxValue error
Now that `array%x` is handled in array expression lowering by genarr,
the code that was packaging the result in the continuations into
an ExtendedValue was now wrong: it was calling
arrayElementToExtendedValue that only works to package the an
element of `array` and `array(i:j:k)` into an Exv given `array`
ExtendedValue. On non-character components, this went unnoticed, but
for character, the length was not propagated, and verifier was
rightfully complaining about "character buffer should be in CharBoxValue".

To fix this, add a arraySectionElementToExtendedValue that takes a slice
argument and deal with the `array(i:j:k)%d%c` generic case.

In the process, it was identified that one usage of
arrayElementToExtendedValue was actually wrong, since it was used to
package an array value (result of ArrayUpdateOp), and not an element.
A similar situation with a usage of fir::substbase that was use to
substitute a scalar exv by an array value. In both case, the resulting
Exv properties are useless, the use place specifically expect an
ArrayUpdateOp result to be place in an ArrayMergeOp. Instead of fighting
to build an Exv with the actual properties for nothing, add
abstractArrayExtendedValue that builds the correct Exv category, but
provides empty values for the properties.