Handle vector subscript in output IO (TODO left for input)
When provided with an expr containing a vector section, genExprBox
was trying to build a fir.box without making a temp. This is not
possible. Fix this by testing for vector subscripts before calling
createSomeArrayBox.

This caused crash: "Assertion `Index < Length && "Invalid index!"' failed."
The crash was mainly observed in IO statement where genExprBox is used.

This fix output IO, but for input IO, something special will need to be done
for array with vector subscript, most likely making a temp, pass that to the
IO input runtime and copy back the temp into the array section. There are
alternatives, so this more complex case is left TODO until a solution
is chosen for it.