Update ArrayUpdateOp to carry both rhs and lhs character length
ArrayValueCopy currently depends on the defining operations of
ArrayUpdateOp merge argument to deduce its length. This is failing
in contexts where merge is coming from a scalar expressions and
it may be impossible to retrieve the dynamic length if it was never
used.

Fix this by allowing ArrayUpdateOp typeparams operand to carry both the
rhs and lhs dynamic length if needed. Add accessors and verifier to deal
with this as safely as possible.

In lowering, centralize fir::ArrayUpdateOp creation and deal with the
character specificity.

In the process, the IterSpace element member has to be made an
ExtendedValue so that the element length can be access when building
ccDest lambda.

The current patch is fixing issue 867 and the issue discovered in
PR887 (lack of padding for constant chars rhs and lhs arrays with
different lengths), but it has not yet been run on regression tests
and lit tests are not yet fixed.