Fix bad locations
PR #865 added a setCurrentLocation in function declaration that had
the side effect of changing some previously "unknown" location into
bad location while lowering functions bodies later.

The root-cause was that lowerFunc itself was not calling setCurrentLocation,
so the first function was now lowered with the location of the last
declaration.

In the process, also set locations for modules, and before starting
lowering do constructs. There are probably other constructs/evaluation
that are not calling setCurrentLocation as they should, but now at least
they won't get the location from a later function.