Fix inquiry intrinsic argument lowering
Inquiry argument lowering was using `evaluate::IsAllocatableOrPointer`
instead of `isAllocatableOrPointer`. The First one does not behave as needed
when lowering inquired arguments, it was returning true on expression `x(i)`
with x and allocatable, when `x(i)` designator should not be considered an
allocatable designator in lowering (it cannot be allocated).