Lower default initialization of local and dummy entities
When instantiating variables, identify the ones that must be default
initialized at runtime (locals and intent(out) dummies). Call the
default initialization runtime routine to perform the initialization.

Global variables that have no explicit initialization and have a default
initialization are left TODO. The runtime cannot be used for them, the
fir.global generation will need to be updated for these variables (it was
generating an undefined undefined until now). For now, simply catch and
add a hard TODO for these cases. For information, it is firing 13 times
in NAG F95 tests.

Runtime interface for finalization is added, but nothing is done with it
yet (Final procedures are not F95).