Apply lower bounds in array component reference
An old FIXME never got fixed. Finally implement it by regrouping the
logic of ArrayRef lowering for both symbols and components base.
Rename the `gen` functions helper to clarify what is happening.

Also force use of collapsed coordinate_of in case the array is a
dynamic length characters without fir.box. Otherwise, the length
info is lost in the fir.coordinate_of.

And prevent collapsing of fir.coordinate_of for rank 1 array with
dynamic length but no dynamic length.

Fix all the tests that were wrong or are impacted by these changes.

Note 1: this fixes the second bug observed in #902.