Lower default initialization of global variables
Add a genDefaultInitializerValue method in variable lowering that
generates the body of the global op that must be default initialized.

Default initialized defers from explicit initialization because all the
component may not have an initialization, therefore the front-end is not
providing a top level ev::Expr that can be lowered for the derived type.

Hence, lowering visits each components and lower any default
initialization if present or leave the component undefined otherwise.

genDefaultInitializerValue is recursive because component without a
default initialization value can be of a derived type whose component
have default initialization.

Default initialization also applies to variable in equivalences. The
standard is not 100% clear about the possibility of mixing overlapping
default and explicit initialization. This PR support full overlaps (the
default initialization is completely overridden by explicit
initialization of other equivalence members), no overlap (the default
initialization is fully apply to the equivalence storage). It however
does not support mixing pieces of explicit initialization from other
equivalence members over part of the default initialization (the
situation is detected and a TODO emitted). Not all compilers support
this last case (nagfor, ifort and nvfortran support this, gfortran refuses it,
and xlf silently ignores the explicit initialization).