[flang] Implement lowering for the SIZE intrinsic
Assumed-rank entities are not yet handled.

No runtime call was needed.  Also, all situations where the DIM argument
is not present are handled by the front end converting calls to SIZE
with values of the DIM argument that correspond to the rank of the ARRAY
argument.  Thus, if ARRAY has rank 1, a call to "SIZE(ARRAY)" is changed
to "SIZE(ARRAY, 1)".  If ARRAY has rank 2, the call is changed to the
expression "SIZE(ARRAY, 1) * SIZE(ARRAY, 2)".  Thus, the only thing I
needed to do was convert the value of the DIM argument to be 1-based and
make the appropriate call to BoxDimsOp.

I also added a test.