Instantiate module variables in internal procedure of module procedures
When an internal procedure is defined inside a module procedure, it may
access the module variables. These access are not materialized by
associated symbols in the internal procedure scope, so lowering must
instantiate all the module variables before lowering the internal
procedure so that any reference to the module symbols can be resolved.

Also update the case where the internal procedure is not in a module
procedure but simply in a procedure that uses module variables. This
case compiled and worked, but the module variable was passed via the
host tuple when it can simply be instantiated directly from the module
address inside the internal procedure.