Ensure scalar args are lowered to ref if needed in elemental calls
Scalars were args were lowered to value in elemental calls where they needed
to be passed by reference. Lower scalar to a reference when the elemental function
takes a reference. In case the elemental arg is passed by value, lower the arg by
value and store the value in a temp that is passed to the call.

This fixes the error "'llvm.call' op operand type mismatch for operand 1" in
nag_f08_95 l7.f90 (similar error remains coming from non elemental context in
some other tests)

It is not 100% clear to me that the by address case is OK from an aliasing point
of view (e.g. in `i = foo(j, i(10))` the current passing the raw address of i(10)
at every iteration).