Lower elemental character function and elemental internal procedure
Split the scalar function lowering genRawProcedureRef in order to share
the part that deals with explicit result allocation and host link
passing with elemental procedure lowering. This code is split without
any functional change.

Use the new function in elemental procedure lowering. Instead of placing
the result manually in a vector or mlir::Value, use the CallerInterface
placeInput function, and let the scalar lowering deal with the result
instantiation and the callOp.
In order to be able to capture the CallerInterface object, its
std::unique_ptr  member is made a std::optional, this enable the
creation of the implicit copy constructor.