[NFC] Move runtime API call generator into Optimizer/Runtime - part 2/2
In order to be able to emit runtime calls in FIR transformation passes,
move the runtime call generators into Optimizer/Runtime.
RTBuilder.h is moved there, as well as DerivedRuntime.h, NumericRuntime.h,
ReductionRuntime.h, CharacterRuntime.h, and TransformationalRuntime.h

All the methods are moved from the Fortran::lower namespace into a new
fir::runtime namespace.

Runtime.[h, cpp] is not moved yet because it uses parse tree nodes.
There is no urgent need to move it now, so I did not attempted to split
it in this PR.

Allocatable.cpp and IO.cpp make direct usages of RTBuilder.h, it may be
cleaner to handle the runtime call generation details in new
Optimizer/xxx files, but this is not done here since this calls are not
currently needed in lib/Optimizer.