Lower derived array assignment and improve derived scalar assignment
Move the genRecordAssignment scalar assignment helper in fir::factory so
that it can be used for the array case in ArrayCopyValue.cpp.

Simplify genRecordAssignment to only lower as inline code the cases where
a mem copy can be done to implement the derive type assignment (that is
when their are no allocatable component, automatic component, or
component requiring a user defined assignment calls). For the rest, use
the runtime genAssign() that handles deep copies and user defined
assignments.

For the inline version, instead of working component by component, and
using the array expression framework for array components, simply emit a
derived type load + store. This is OK since the derived types assigned
inlined have constant size. LLVM transform this into element by element
assignment of llvm.memmove itself (opt --memcpyopt inserts memmove calls).

This allows implementing the simple cases inlined without depending on
the array expression framework (which is not possible during the
ArrayCopyValue pass), or manually creating llvm.memmove, which would
require inserting a fir.sizeof of to compute the byte size for the
memmove.