Zero size arrays of derived type (#1050)
Fix two problems with zero size arrays with derived type elements.

The first problem occurs when the derived type element of a zero size
array has default initialization.  -- fix in ConvertVariable.cpp

  type dt
    integer :: j = 17
  end type
  type(dt) :: z(0)

The second problem occurs when the derived type element of a zero size
array has an explicit, conformable, scalar initialization.  That appears
to be standard conforming, if not too useful, although it has the side
affect of forcing the SAVE attribute on z.  -- fix in ConvertExpr.cpp

  type dt
    integer :: j
  end type
  type(dt) :: z(0) = dt(99)

Also add a verifier for fir.insert_on_range ops, which triggers for
the first test case without the corresponding fix.