Ensure temporaries are passed in calls for parenthesized arguments
foo(x) and foo((x)) are semantically different. In the former, the dummy
argument is associated with x storage, while in the latter, the dummy
argument is associated with a temp initialized with x value.

This is an important distinction in calls like foo(x, (x)) where passing
x storage twice would be wrong, also to not propagate back potential changes
on the dummy.

Parentheses are not lowered to temps in lowering because we do not want
to create temp array sub-expression. Only a fir.no_reassoc is created
for parentheses by expression lowering.

This change keep this general behavior and add special handling in call
lowering to ensure a temp is created for parenthesized arguments.